requirements.txt.txt
Flask
pymongo
psycopg2-binary
azure-storage-blob
azure-storage-queue
gunicorn